Replaced the test subject with long
When converting to MethodHandle, Long is assumed to be given priority.

+++ src/test/kotlin/io/github/projectmapk/jackson/module/kogera/zIntegration/deser/valueClass/ValueClasses.kt
@@ -18,13 +18,13 @@ value class Primitive(val v: Int) {
 }
 
 @JvmInline
-value class TwoUnitPrimitive(val v: Double) {
+value class TwoUnitPrimitive(val v: Long) {
     class Deserializer : StdDeserializer<TwoUnitPrimitive>(TwoUnitPrimitive::class.java) {
-        override fun deserialize(p: JsonParser, ctxt: DeserializationContext): TwoUnitPrimitive = TwoUnitPrimitive(p.doubleValue + 100)
+        override fun deserialize(p: JsonParser, ctxt: DeserializationContext): TwoUnitPrimitive = TwoUnitPrimitive(p.longValue + 100)
     }
 
     class KeyDeserializer : JacksonKeyDeserializer() {
-        override fun deserializeKey(key: String, ctxt: DeserializationContext) = TwoUnitPrimitive(key.toDouble() + 100)
+        override fun deserializeKey(key: String, ctxt: DeserializationContext) = TwoUnitPrimitive(key.toLong() + 100)
     }
 }
