+++ src/test/kotlin/io/github/projectmapk/jackson/module/kogera/zIntegration/deser/valueClass/parameterSize/twoUnitPrimitive/DeserializeByFactoryWithoutDefaultArgumentsTest.kt
@@ -128,7 +128,7 @@ class DeserializeByFactoryWithoutDefaultArgumentsTest {
 
     @Test
     fun test32() {
-        val expected = callPrimaryConstructor<Dst32> { TwoUnitPrimitive(it.index.toDouble()) }
+        val expected = callPrimaryConstructor<Dst32> { TwoUnitPrimitive(it.index.toLong()) }
         val src = replacePQ(defaultMapper.writeValueAsString(expected))
         Assertions.assertEquals(expected, defaultMapper.readValue<Dst32>(src))
     }
@@ -245,7 +245,7 @@ class DeserializeByFactoryWithoutDefaultArgumentsTest {
 
     @Test
     fun test33() {
-        val expected = callPrimaryConstructor<Dst33> { TwoUnitPrimitive(it.index.toDouble()) }
+        val expected = callPrimaryConstructor<Dst33> { TwoUnitPrimitive(it.index.toLong()) }
         val src = replacePQ(defaultMapper.writeValueAsString(expected))
         Assertions.assertEquals(expected, defaultMapper.readValue<Dst33>(src))
     }
@@ -455,7 +455,7 @@ class DeserializeByFactoryWithoutDefaultArgumentsTest {
 
     @Test
     fun test64() {
-        val expected = callPrimaryConstructor<Dst64> { TwoUnitPrimitive(it.index.toDouble()) }
+        val expected = callPrimaryConstructor<Dst64> { TwoUnitPrimitive(it.index.toLong()) }
         val src = replacePQ(defaultMapper.writeValueAsString(expected))
         Assertions.assertEquals(expected, defaultMapper.readValue<Dst64>(src))
     }
@@ -668,7 +668,7 @@ class DeserializeByFactoryWithoutDefaultArgumentsTest {
 
     @Test
     fun test65() {
-        val expected = callPrimaryConstructor<Dst65> { TwoUnitPrimitive(it.index.toDouble()) }
+        val expected = callPrimaryConstructor<Dst65> { TwoUnitPrimitive(it.index.toLong()) }
         val src = replacePQ(defaultMapper.writeValueAsString(expected))
         Assertions.assertEquals(expected, defaultMapper.readValue<Dst65>(src))
     }
@@ -1065,7 +1065,7 @@ class DeserializeByFactoryWithoutDefaultArgumentsTest {
 
     @Test
     fun testMax() {
-        val expected = callPrimaryConstructor<DstMax> { TwoUnitPrimitive(it.index.toDouble()) }
+        val expected = callPrimaryConstructor<DstMax> { TwoUnitPrimitive(it.index.toLong()) }
         val src = replacePQ(defaultMapper.writeValueAsString(expected))
         assertReflectEquals(expected, defaultMapper.readValue<DstMax>(src))
     }
