+++ src/test/kotlin/io/github/projectmapk/jackson/module/kogera/zIntegration/ser/valueClass/jsonInclude/ValueClasses.kt
@@ -10,4 +10,4 @@ value class NonNullObject(val v: String)
 value class NullableObject(val v: String?)
 
 @JvmInline
-value class TwoUnitPrimitive(val v: Double)
+value class TwoUnitPrimitive(val v: Long)
