+++ src/test/kotlin/io/github/projectmapk/jackson/module/kogera/zIntegration/ser/valueClass/jsonKUnbox/ForClass.kt
@@ -20,7 +20,7 @@ class ForClass {
 
     @JsonKUnbox
     @JvmInline
-    value class TwoUnitPrimitive(val v: Double)
+    value class TwoUnitPrimitive(val v: Long)
 
     data class Dto(
         val p0: Primitive = Primitive(0),
@@ -33,15 +33,15 @@ class ForClass {
         val no1: NullableObject = NullableObject(null),
         val no2: NullableObject? = NullableObject("2"),
         val no3: NullableObject? = null,
-        val tup0: TwoUnitPrimitive = TwoUnitPrimitive(0.0),
-        val tup1: TwoUnitPrimitive? = TwoUnitPrimitive(1.0),
+        val tup0: TwoUnitPrimitive = TwoUnitPrimitive(0),
+        val tup1: TwoUnitPrimitive? = TwoUnitPrimitive(1),
         val tup2: TwoUnitPrimitive? = null
     )
 
     @Test
     fun test() {
         val expected = """
-            {"p0":0,"p1":1,"p2":null,"nno0":"0","nno1":"1","nno2":null,"no0":"0","no1":null,"no2":"2","no3":null,"tup0":0.0,"tup1":1.0,"tup2":null}
+            {"p0":0,"p1":1,"p2":null,"nno0":"0","nno1":"1","nno2":null,"no0":"0","no1":null,"no2":"2","no3":null,"tup0":0,"tup1":1,"tup2":null}
         """.trimIndent()
         val actual = jacksonObjectMapper().writeValueAsString(Dto())
