+++ src/test/kotlin/io/github/projectmapk/jackson/module/kogera/zIntegration/ser/valueClass/jsonKUnbox/ForProperty.kt
@@ -16,7 +16,7 @@ class ForProperty {
     value class NullableObject(val v: String?)
 
     @JvmInline
-    value class TwoUnitPrimitive(val v: Double)
+    value class TwoUnitPrimitive(val v: Long)
 
     data class Dto(
         @get:JsonKUnbox
@@ -40,9 +40,9 @@ class ForProperty {
         @get:JsonKUnbox
         val no3: NullableObject? = null,
         @get:JsonKUnbox
-        val tup0: TwoUnitPrimitive = TwoUnitPrimitive(0.0),
+        val tup0: TwoUnitPrimitive = TwoUnitPrimitive(0),
         @get:JsonKUnbox
-        val tup1: TwoUnitPrimitive? = TwoUnitPrimitive(1.0),
+        val tup1: TwoUnitPrimitive? = TwoUnitPrimitive(1),
         @get:JsonKUnbox
         val tup2: TwoUnitPrimitive? = null
     )
@@ -50,7 +50,7 @@ class ForProperty {
     @Test
     fun test() {
         val expected = """
-            {"p0":0,"p1":1,"p2":null,"nno0":"0","nno1":"1","nno2":null,"no0":"0","no1":null,"no2":"2","no3":null,"tup0":0.0,"tup1":1.0,"tup2":null}
+            {"p0":0,"p1":1,"p2":null,"nno0":"0","nno1":"1","nno2":null,"no0":"0","no1":null,"no2":"2","no3":null,"tup0":0,"tup1":1,"tup2":null}
         """.trimIndent()
         val actual = jacksonObjectMapper().writeValueAsString(Dto())
